[2417] 정수 제곱근

+++ src/Main.java
@@ -5,6 +5,6 @@ class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         long n = sc.nextLong();
-        System.out.println((int) Math.sqrt(n));
+        System.out.println((long)(Math.sqrt(n)));
     }
 }
